class wc-alcohol
